instability mechanic
Added an Instability mechanic to the crafting altar to spice things up, not entirely unlike TC but not the same either. The more components a spell has, the more instability it has; stacking modifiers increases instability. Having more than one player nearby while crafting a spell increases instability; crafting in the middle of a storm increases instability; Having a black aurem nearby increases instability. To increase stability, make the altar out of better blocks; To counter instability, you can also surround the altar with wizards chalk and liquid etherium. Having a celestial prism nearby helps. Finally, performing the crafting at night and having the altar be open to the sky slightly decreases instability. The instability check is performed every time a component is added, but is random, so you might get lucky even with little stability. If it fails, you have a chance to: be drained of your mana; set your surroundings on fire; attract a lightning; explode (strength of explosion depends on amount of instability); suffer damage; If the instability is sufficiently high, you will also attact arbitrary types of Elementals. Finally, if you go out of your way to have extremely high instability, you have a chance to be teleported to an arbitrary point in spacetime.