Added Entropic Enervator
This block allows to, on purpose, controllably break items. It can turn stone bricks to cracked stone bricks, stone to cobblestone to gravel, bones to bonemeal, cobweb to strings, sandstone to sand, and any music disks to the broken music disk (C418 - 11). Although, be wary: upon placing any tool with durability inside, it will decrease the durability of said tool.