Added my vision of Archmage Armor
This is an endgame set of armor, to compete with Ichor and Bound/Omega armor. The crafting recipes are extremely difficult. When wearing the Archmage Hood: you are granted fire resistance and night vision; Creatures that hit you are set on fire. When wearing the Archmage Chestplate: You are granted knockback resistance. Your hits knock enemies back. You are granted 2 extra rows of hearts. When wearing the Archmage Leggings: You are granted water breathing; You can move in water without difficulty; Breaking blocks underwater is easier. When wearing the Archmage Boots: You are granted protection from fall damage. You are not slowed down by Slowness. You are granted flight. When wearing the full set: Mana regeneration is accelerated; You are granted the ability to 'Lash' creatures to the sky, reversing their gravity for a short period of time, by hitting them while sneaking.