Contingency Changes
Made it so one contingency of each type can be activate at the same time, instead of one contingency in total; Made damage contingencies (damage and health) be able to be applied to the attacker, if the damage came from an entity; Added Bound Armor (not to be confused with blood magic's Bound Armor). Bound Armor is worse than battlemage/archmage in terms of protection and doesn't offer any mana regen or buffs on its own, however: you can bind spells to it using the crafting table. The spells will then have a contingency added to them (do not include the contingency in the spell itself) (helmet - death, chest - health, legs - damage, boots - fall), and will renew it every few seconds on the condition that you do not already have a contingency of that type. This does not grant a gameplay advantage other than the convenience of having your contingencies auto-cast themselves.