Attepted fixes
With the summoning, I made it so summoned entities will disappear outside of a certain range if you attempt to summon another one, as there is supposedly an issue with summons and chunk unloading where it isn't done correctly. With the power nodes, I'm more shooting in the dark. This will probably not solve anything, but it's worth a shot anyway. If you can repeatedly reproduce the mana-battery-charge-disappearing issue, please let me know.