Added Lingering Modifier
Essentially, this modifiers makes a spell persist. Best to explain with examples: let's imagine a spell 'projectile:ignition:lingering'. It will ignite a block, but if you put it out within roughly half a second, if will ignite the block again. If there are multiple Lingering modifiers, it will do this multiple times. Duration modifiers can increase the delay between persistence updates. As a separate feature, applying this modifier to a 'Zone' shape and casting it on an entity will have the effect of the zone following the entity for some time, creating a Cloak-like effect.